+++ database/factories/NetworkBannerFactory.php
@@ -33,6 +33,7 @@ class NetworkBannerFactory extends Factory
     {
         return [
             'uuid' => $this->faker->uuid,
+            'demand_banner_id' => $this->faker->uuid,
             'network_campaign_id' => $this->faker->randomDigit(),
             'serve_url' => $this->faker->url,
             'view_url' => $this->faker->url,